Partially complete damage
Parenting/unparenting of enemies on arm required

+++ Assets/Scripts/Alejandro/Enemies/EnemyBase.cs
@@ -6,7 +6,7 @@ public abstract class EnemyBase : MonoBehaviour
 {
     [SerializeField] SO_EnemySettings settings;
 
-    protected int currentHealth;
+    protected float currentHealth;
     protected float moveSpeed;
 
     [SerializeField] protected List<Transform> internalWaypoints;
